pulsar-client-python-52/requirements.txt
Flask==2.0.1
gevent==21.8.0
gunicorn==20.1.0
pulsar-client==2.10.2
